Fix copyit-pandoc commands and documented it
resolve https://github.com/zonuexe/emacs-copyit/issues/2

The usage of these commands was not consistent with other copyit
commands.

From now on, as with the other copyit commands, a file is selected by
default, and the contents of the current buffer are converted when
prefixed with `C-u`.

Thank you @xbelanch